+++ HISTORY.md
@@ -1,6 +1,6 @@
 # History
 
-## 0.5.0 (2021-TBD)
+## 0.5.0 (2021-09-TBD)
 [See all issues & PRs for v0.5](https://github.com/JWCook/aiohttp-client-cache/milestone/4?closed=1)
 
 * Add a filesystem backend
@@ -25,7 +25,7 @@
 * Fix initialziation of `SQLiteBackend` so it can be created outside main event loop
 
 ## 0.4.0 (2021-05-12)
-[See all issues & PRs for v0.3](https://github.com/JWCook/aiohttp-client-cache/milestone/3?closed=1)
+[See all issues & PRs for v0.4](https://github.com/JWCook/aiohttp-client-cache/milestone/3?closed=1)
 
 * Add optional support for the following **request** headers:
     * `Cache-Control: max-age`